requirements.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==1.9.0
django-allauth==0.37.1
django-filer==1.3.2
django-mptt==0.8.7
django-polymorphic==1.3.1
django-rest-authemail==1.4.0
django-simple-history==1.9.1
djangorestframework==3.3.3
easy-thumbnails==2.5
idna==2.7
oauthlib==2.1.0
Pillow==5.3.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
Unidecode==0.4.21
urllib3==1.23
